Refactor email
